Move Partwise.Part and Partwise.Measure to own files
